Updated component name
